refactor(vm2): split bytecode helpers
